requirements.txt
pyperclip
pyaudio
python-ulid